pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "resume-forge-mcp"
version = "0.4.0"
description = "MCP server for intelligent LaTeX resume generation — browse 350+ Overleaf templates, auto-compile with pdflatex/xelatex/lualatex, tailor to job descriptions, ATS scoring, and visual preview."
readme = "README.md"
license = "MIT"
authors = [
    { name = "Jitendra Singh Sankhwar" },
    { name = "Danny Liu" }
]
keywords = ["mcp", "latex", "resume", "cv", "claude", "ai", "overleaf", "pdf"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
preview = ["pymupdf>=1.24.0"]
pdf = ["pymupdf>=1.24.0"]
docx = ["python-docx>=1.0.0"]
all = ["pymupdf>=1.24.0", "python-docx>=1.0.0"]

[project.scripts]
resume-forge-mcp = "resume_forge_mcp:main"

[project.urls]
Homepage = "https://github.com/jitendrasinghsankhwar/resume-forge-mcp"
Repository = "https://github.com/jitendrasinghsankhwar/resume-forge-mcp"
Issues = "https://github.com/jitendrasinghsankhwar/resume-forge-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/resume_forge_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["W191", "E101"]

[tool.ruff.format]
indent-style = "tab"
quote-style = "double"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.bandit]
exclude_dirs = ["tests"]
